Renamed conflicting folder names and fixing this across the app.  Added the action to return the guid result payload on receiving a post request.
