+++ Sulfur/Services/UrlPayloadActions/IUrlPayloadService.cs
@@ -4,11 +4,11 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Sulfur.Services.UrlPayload
+namespace Sulfur.Services.UrlPayloadActions
 {
     //Methods to be used by the UserPayloadService
     public interface IUrlPayloadService
     {
-        GuidResult GenerateGuidPayload(string url);
+        GuidResult GenerateGuidPayload();
     }
 }
